% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cochran.R
\name{cochran_test}
\alias{cochran_test}
\title{cochrans q test}
\usage{
cochran_test(
  .data,
  formula,
  type = c("exact", "montecarlo"),
  n_samples = 10000
)
}
\arguments{
\item{.data}{dataframe}

\item{formula}{defines block and group variables of the order lhs ~ group | block}

\item{type}{how to compute p? exact or montecarlo}
}
\value{
list with statistic and corresponding one sided p value
}
\description{
cochrans q test
}
\examples{
autos <- matrix(data = c(0,0,0,0,0,1,0,0,0,1,
                         0,1,0,0,1,0,0,1,0,1,
                         0,1,1,0,1,1,0,1,1,1,
                         0,1,1,0,0,1,1,1,1,1),
                nrow = 10,
                ncol = 4,
                byrow = FALSE) \%>\%
                as_tibble() \%>\%
                mutate(family = row_number()) \%>\%
                pivot_longer(V1:V4, names_to = "group")
cochran_test(value ~ group | family,
              formula = value ~ group | family,
              type = "montecarlo")
}
