% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedman.R
\name{convert_base}
\alias{convert_base}
\title{base converter}
\usage{
convert_base(num, base)
}
\arguments{
\item{num}{number to convert}

\item{p}{base to convert to}
}
\value{
vector of decimal values
}
\description{
base converter
}
\examples{
convert_base(22,4)
}
