% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cochran.R
\name{create_q_f}
\alias{create_q_f}
\title{create a compute q function}
\usage{
create_q_f(.data, formula)
}
\arguments{
\item{.data}{dataset (may be permuted)}

\item{formula}{defines block, dv, and group variable}
}
\value{
q test statistic
}
\description{
create a compute q function
}
\examples{
autos <- matrix(data = c(0,0,0,0,0,1,0,0,0,1,
                         0,1,0,0,1,0,0,1,0,1,
                         0,1,1,0,1,1,0,1,1,1,
                         0,1,1,0,0,1,1,1,1,1),
                nrow = 10,
                ncol = 4,
                byrow = FALSE) \%>\%
                as_tibble() \%>\%
                mutate(family = row_number()) \%>\%
                pivot_longer(V1:V4, names_to = "group")
create_q_f(value ~ group | family, formula = value ~ group | family)
}
