% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedman.R
\name{friedman_test}
\alias{friedman_test}
\title{compute friedman statistic and p value}
\usage{
friedman_test(
  .data,
  formula,
  type = c("exact", "montecarlo"),
  n_samples = 10000
)
}
\arguments{
\item{.data}{dataframe including all relevant vectors}

\item{formula}{of the form: dv ~ group | block}

\item{type}{exact}
}
\value{
list with p-value and test statistic of Friedmans test
}
\description{
compute friedman statistic and p value
}
\examples{
participant <- 1:4
group <- 1:3
data <- expand.grid(
  "participant" = participant,
  "group" = group
)
data$ability <- c(
  7, 8, 8, 6,
  10, 12, 8, 10,
  11, 12, 13, 11
)


friedman_test(.data = data, formula = ability ~ group | participant)
}
