% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedman.R
\name{get_q_light}
\alias{get_q_light}
\title{light version of get_q
does less computation
stuff that can be computed once is computed outside the function}
\usage{
get_q_light(.data, formula, total_t, np)
}
\arguments{
\item{np}{}
}
\description{
light version of get_q
does less computation
stuff that can be computed once is computed outside the function
}
