% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedman.R
\name{get_tie_correction}
\alias{get_tie_correction}
\title{compute tie correction}
\usage{
get_tie_correction(.data, vector)
}
\arguments{
\item{.data}{dataframe, may be grouped}

\item{vector}{name of vector in dataframe on which you want to compute the tie vector}
}
\value{
dataframe including t as tievector
}
\description{
compute tie correction
}
\examples{
get_tie_correction(.data = tibble(vec = c(1:6, 1:6)), vec)
}
