% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedman.R
\name{nth_row}
\alias{nth_row}
\title{nth row
get nth row of permutation table of certain specification.
namely, the p permutation vectors are all the same and are consecutive integers
starting form 1 to p}
\usage{
nth_row(n, p = 3, ninp = 5)
}
\arguments{
\item{n}{numeric(1) row of permutation table}

\item{p}{numeric(1) number of variables}

\item{ninp}{numeric(1) length of original vector in each variable}
}
\value{
indicator vector with element for each block to get order from permutation list
}
\description{
nth row
get nth row of permutation table of certain specification.
namely, the p permutation vectors are all the same and are consecutive integers
starting form 1 to p
}
\examples{
lapply(1:1000, function(n) nth_row(n, 30, 5))
# or simply:
nth_row(10)
# basevec <- rep(1, p)
# x <- Rmpfr::mpfr(n - 1, precBits = 64, base = 10) ## base = 10 is default
# numb <- as.numeric(Rmpfr::formatMpfr(x, base = ninp))
# vec <- as.numeric(
#   strsplit(as.character(numb), "") \%>\% purrr::as_vector()
# )
# vec <- c(rep(0, length(basevec) - length(vec)), vec)
# basevec + vec
}
